made structure better
