+++ app/config.py
@@ -5,3 +5,11 @@ load_dotenv()
 
 GROQ_API_KEY = os.getenv("GROQ_API_KEY")
 DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./app.db")
+
+TWILIO_ACCOUNT_SID = os.getenv("TWILIO_ACCOUNT_SID")
+TWILIO_API_KEY = os.getenv("TWILIO_API_KEY")
+TWILIO_API_SECRET = os.getenv("TWILIO_API_SECRET")
+TWIML_APP_SID = os.getenv("TWIML_APP_SID")
+
+# domyślny użytkownik backendu (na razie na sztywno)
+BACKEND_USER_ID = os.getenv("BACKEND_USER_ID", "user-1")
