--- heavy_requirements.txt
@@ -1,3 +1,0 @@
-speechrecognition
-pyttsx3
-pyaudio
