+++ requirements.txt
@@ -6,3 +6,4 @@ python-dotenv
 groq
 httpx
 python-multipart
+twilio
